The syntax given in the Help file for Calc’s LOGNORM.INV function is incorrect (see Bug 137649). Specifically, the second (“Mean”) and third (“StDev”) arguments are optional rather than required in Calc’s current implementation. If omitted the Mean argument defaults to 0.0 and the StDev argument defaults to 1.0.
Note: The corresponding Excel function’s three arguments are all required (none are optional) and so one possible outcome of Bug 137649 is for Calc to be updated to match this behaviour for interoperability. However, we do not know if or when this might happen and so, in the meantime, I propose to update the Help to more accurately reflect Calc’s current behaviour.

The change affects only one file, 04060183.xhp.

Change-Id: I69934a8123d5de41007bc4dfca1cbc478ed963a6
Reviewed-on: https://gerrit.libreoffice.org/c/help/+/110150
Tested-by: Jenkins
Reviewed-by: Olivier Hallot <olivier.hallot@libreoffice.org>