Update patch set 2
Patch Set 2:

Although I am only a cc recipient, I have a couple of minor comments. I trust that’s okay, and I hope these thoughts are helpful.

(1)	In the introductory paragraph is the text “If the optional parameter NoThousandsSeparators is TRUE”. I would suggest changing “TRUE” to “non-zero” to match the later description of the parameter.

(2)	Following up on Rafael’s comment about the text “If Decimals is a fraction, it is rounded towards 0 as an integer (ignoring what is the closest integer).”. My experimentation seems to indicate that FIXED applies a floor function (so rounds +ve numbers down towards zero, and -ve numbers down away from zero).

(3)	The second example appears to contain a spurious close round bracket.

(4)	Is it worth mentioning that -15 <= Decimals <= 15?

Patch-set: 2