Help Pages for Calc’s “IM” Functions, Improved Equations
I have recently spent some time updating the pages
of the Calc Functions Wiki that relate to the
processing of complex numbers.

There are 25 functions in Calc’s Add-In category
that have names beginning with the letters “IM”
and these are all concerned with processing
complex numbers of the form “a+bi”.

Of these 25, there are 6 trigonometric functions
(IMCOS, IMCOT, IMCSC, IMSEC, IMSIN, and IMTAN)
and 4 hyperbolic functions (IMCOSH, IMCSCH,
IMSECH, and IMSINH). For these 10, the current
help pages include equations to assist the
user’s understanding of the function’s purpose.

I feel that some of these equations should be
replaced with more appropriate equations that
reflect how the software functions (see
/core/scaddins/source/analysis/analysishelper.cxx)
and these proposed changes are listed below.

Change-Id: Ie0e4957e374570d60be00bcbe9b9015acab5e3a1
Reviewed-on: https://gerrit.libreoffice.org/c/help/+/113935
Tested-by: Jenkins
Reviewed-by: Olivier Hallot <olivier.hallot@libreoffice.org>