Fix sort command in Help pages
After the introduction of the headings h1 ... h6  in the Help
pages the sort command broke. It relied only in the <paragraph>
tag for sorting, usually the first paragraph in the section.

The patch extend the DTD to handle the new descendant="" attribute
to select which tag to sort in the Help page

Default is <paragraph> for backward compatibility.

Pages with <sort> must be revisited.

Change-Id: I3c6f243369429ed7824a68b11757dc15f4555c0f