Update patch set 4
Patch Set 4:

Thanks Rafael.

With regards to point 1, I do not think that the current Calc behaviour is ideal or consistent – for other functions you can access the help from the Function Wizard in a context-sensitive way. However, I do not think that is an issue to be addressed through this change so best forget that I mentioned it!

With regards to point 2. Can I suggest changing the third sentence of the first paragraph to “The units of measurement specified through the arguments must match the supported unit symbols, which are case-sensitive”?

With regards to point 5. I remain uncomfortable with the inclusion of this link when I think about consistency with the other 500+ Calc functions! Does Olivier have a view on it?

Otherwise, fine.

Patch-set: 4
Reviewer: Gerrit User 1002588 <1002588@6622c302-437a-49ad-a91d-55ba7dcf1615>