Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> Thanks Rafael.
> 
> With regards to point 1, I do not think that the current Calc behaviour is ideal or consistent – for other functions you can access the help from the Function Wizard in a context-sensitive way. However, I do not think that is an issue to be addressed through this change so best forget that I mentioned it!

Actually, there is a recurring bug in the Help and (many) dialogs. If you press F1 on a function in the Function wizard, you will be sent to the right Help page. The bug is on the Help button. This is becasue the Help button belongs to the "dialog frame" and not the internal function content. A bug should be open for the Function Wizard, CC: Caolán Mc Namara. 

> 
> With regards to point 2. Can I suggest changing the third sentence of the first paragraph to “The units of measurement specified through the arguments must match the supported unit symbols, which are case-sensitive”?
> 
> With regards to point 5. I remain uncomfortable with the inclusion of this link when I think about consistency with the other 500+ Calc functions! Does Olivier have a view on it?

I'm OK to add a link to the wiki page. However I'll put as a "Related topic" entry in the bottom of the page.

No need to rush to add the link to every page, let's do it slowly, upon reviewing other topics on the same file.

> 
> Otherwise, fine.

Patch-set: 4
Reviewer: Gerrit User 1000059 <1000059@6622c302-437a-49ad-a91d-55ba7dcf1615>