tdf#102523 Improve Advanced Filter help page
This patch embeds the page specialfilter.xhp and 12040201.xhp into the 12040300.xhp file, which is the landing page when the user clicks the Help button in the Advanced Filter dialog.

As a result, the landing page now covers all the features in the dialog and also presents a useful example.

This also updates the .tree file.

Change-Id: I25d31838fa2604851962fdf8200b887255489017
Reviewed-on: https://gerrit.libreoffice.org/c/help/+/118191
Tested-by: Jenkins
Reviewed-by: Olivier Hallot <olivier.hallot@libreoffice.org>