Update patch set 2
Patch Set 2: Code-Review-1

Thanks for your efforts on this update Rafael. A couple of very minor comments:

(1) I’m not sure that it is of any consequence, but the id attribute of the topic tag ("calcfuncregex") is inappropriate.

(2) Fourth example. Can I suggest a slight re-word, replacing “makes the numeric part be” with the alternative “causes the numeric part to be”?

Patch-set: 2
Reviewer: Gerrit User 1002588 <1002588@6622c302-437a-49ad-a91d-55ba7dcf1615>
Label: Code-Review=-1