Update patch set 3
Patch Set 3:

(5 comments)

Hi Olivier. As always, thanks for your feedback.

As for XML refactoring, in future patches I will separate them in two: (i) the first to refactor the XML file and (ii) the second to actually change the contents.

For this patch I think it's best to leave as it is, since you have already made your comments to the patch. Is that ok?

Patch-set: 3