tdf#123506 - HTML Help pages - group search results under module headings
The "double entries" reported are because the search results
are never grouped according to module (CALC, WRITER, IMPRESS, etc).
Results are now grouped.

The HTML Help page search function uses a "fuzzy" algorithm to pick
matching results. This observation by the reporter is not a bug and
so the main Help page has been updated to include an explanation.

The main Help page has also been updated to draw attention to the
highlighing of the matching characters from the search term, to show
exactly how the match was made.

Change-Id: Ibdb9ac726b36b9dd6f605c854192ec644bfe09e1