tdf#33749 - HTML Help pages - reflect changes to cell merge in Calc
Replaced the entire "Format > Merge Cells" submenu with the single
"Format > Merge and Center Cells" button with all the same behaviour
it has on the Formatting Bar.

Replaced the "Split Cells" button from the context menu of a selected
merged cell with the "Merge and Center Cells" button showing the
current merge status, as it does elsewhere.

Also, add the "Merge and Center Cells" link to the Formatting Bar
Description Help page. It was missing.

Change-Id: I8f200faa7a218adae533752def032f4b7f28eed6