Switch from Fuzzysort to FlexSearch
FlexSearch has the best performance of JS search libs at the moment.
It doesn't have fuzzy search yet, but maybe our users don't even
want the feature. RFE is https://github.com/nextapps-de/flexsearch/issues/118

flexsearch.debug.js is used because it is the only unminified version and
we can't use minified files in the source code as they are equivalent to
binary files (close source). There doesn't seem to be anything that would
make it a bad idea to use the debug version in production, only a couple
of edge case checks that print warnings to console.

This commit also makes automatic filtering work when the input field already
has a string upon page load (functions were called in the wrong order).

Change-Id: I5055f7f7f99cad92dfa3a544e5314eb85a1faced
Reviewed-on: https://gerrit.libreoffice.org/c/help/+/122566
Tested-by: Jenkins
Tested-by: Olivier Hallot <olivier.hallot@libreoffice.org>
Reviewed-by: Rafael Lima <rafael.palma.lima@gmail.com>
Reviewed-by: Olivier Hallot <olivier.hallot@libreoffice.org>