tdf#123506 - focus search results to user's current module
This change focuses the search results on the module the user initially
called help from and in addition, in the same way that the initial index
is presented, limits the results to that module plus GLOBAL results.

Explanation:
Currently, when a user arrives at a help page from a module they begin
with a blank search term and by default the index is focused on the module
they came from.

As they enter a search term the results from ALL modules are presented,
paginated and either, (7.3 current) grouped by modules but in a fixed
order of modules that often pushes results relative to the user's
module off the first page of the index or, (7.2 and earlier)
pseudo randomly with incorrect grouping under modules.

If a user wishes to explor other module results they can choose a module
from the Modules drop down. Their search term remains and they can
easily search again.

Change-Id: Ib10dd55d1c68a839938842fe3607a7a7619b3538
Reviewed-on: https://gerrit.libreoffice.org/c/help/+/122903
Tested-by: Jenkins
Reviewed-by: Olivier Hallot <olivier.hallot@libreoffice.org>
Reviewed-by: Rafael Lima <rafael.palma.lima@gmail.com>