Update help description of Calc’s TRUNC function.
The help description for Calc’s TRUNC function currently does not indicate that the Count argument is optional. However, the help description for the ROUNDDOWN function does show the Count argument as optional.

Since both user functions are implemented via one C++ function (ScRoundDown()), the optionality status of arguments should be the same for both functions. This was checked by confirming that the formulas =ROUNDDOWN(-45.67) and =TRUNC(-45.67) give the same return value (-45).

This change updates the optionality status of Count for the TRUNC function and provides the additional example given above.

Change-Id: I2078d068b870dd05f4c977a2f9909537ecf7f9ab
Reviewed-on: https://gerrit.libreoffice.org/c/help/+/124190
Tested-by: Jenkins
Reviewed-by: Olivier Hallot <olivier.hallot@libreoffice.org>