Improve Example section of help description for Calc's SUM() function
The ability to use SUM in an array formula is important but the help description of an example array formula is uncharacteristically long. Now that the wiki has a more specific example, it is appropriate to shorten the Example section of this help topic.

Also moved SUM description to its own XHP file.

Change-Id: Ic062686c53a1ae8b81100e2b49781aaa79e6ee50