Refactor help tree files
This patch applies "xmlling --format" to all *.tree files in helpcontent2/source/auxiliary.

The reason for this refactoring is that these files' indentation is very messy and it makes it harder to maintain and update the files. Also, the indentation used here relies on Tabs whereas XHP files use two blank spaces as indentation.

So this refactoring also applies 2 blank spaces as indentation to the tree files.

Change-Id: I54366d6efd011a09baf45c3b74c089bd9b0401f1
Reviewed-on: https://gerrit.libreoffice.org/c/help/+/125718
Tested-by: Jenkins
Reviewed-by: Olivier Hallot <olivier.hallot@libreoffice.org>