tdf#144133 Fix help target for Expert Config Edit dialog
When the user opens the "Expert Configuration" dialog, selects a preference and double clicks it, another dialog is opened to edit the preference value. Clicking "Help" in this second dialog leads to a 404 page not found error.

This patch changes this behavior so that clicking the Help button opens the "Expert Configuration" help page.

Change-Id: I0104617153f710b09c7436b7f5e3faa78cddb2bd