Drop reference to AskBot and update link
It is never a good idea to hardcode the name of the backend software for
a website in a translatable string directed to users. It’s not future-proof.
Also, we’ve never called it that; it’s always been “Ask LibreOffice”.

But in this case, I’ve carefully avoided mentioning the product name, for
the sake of corporate builds.

Change-Id: Ic4c1c01e7ecc86d44157fec638a5d79dcb809733