tdf#149353 positioning objects in multipage and multicolumn paragraphs
Presented as <note> in "to paragraph" (shared/01/05260200.xhp),
while only the text is embedded as a paragraph in "to character"
(shared/01/05260300.xhp) because there is already a <tip> on that
page, without much other text.

<note> added to swriter/guide/anchor_object.xhp that positioning
is only on the page where the anchor is located.

Change-Id: I8173050620fbeb8fc78d747ad05d8ab1f74f1346
Reviewed-on: https://gerrit.libreoffice.org/c/help/+/135169
Tested-by: Jenkins
Reviewed-by: Seth Chaiklin <sdc.blanco@youmail.dk>
(cherry picked from commit 47bbc1f86ff89ea5b7f6908fc7b1a1b63cb48ca1)