tdf#149353 positioning objects in multipage and multicolumn paragraphs
Presented as <note> in "to paragraph" (shared/01/05260200.xhp),
while only the text is embedded as a paragraph in "to character"
(shared/01/05260300.xhp) because there is already a <tip> on that
page, without much other text.

<note> added to swriter/guide/anchor_object.xhp that positioning
is only on the page where the anchor is located.

////
<note> Objects can only be positioned on the page where their anchor is located.</note>

Positioning objects in paragraphs that extend over two pages
is restricted to the page where the anchor is located.
For "to paragraph" anchoring, if a paragraph extends over two pages, then
the positioning controls only use the portion of the paragraph on the page
where the anchor is located.


<note id="par_id371654784167792"> Objects can only be positioned on the page where their anchor is located.</note>
to swriter/guide/anchor_object // after table.

<note id="par_id751654784158415"><variable id="paraposition">When a paragraph is used for positioning and the paragraph extends over more than one page or column, then positioning options that refer to a <menuitem>paragraph area</menuitem> or <menuitem>paragraph text area</menuitem> are applied only to the portion of the paragraph in the page or column where the anchor is located</variable>.</note>
<paragraph role="paragraph" id="par_id491654784636830"><embedvar href="text/shared/01/05260200#paraposition"/></paragraph>


swriter/guide/anchor_object

Where should this be placed?! text/shared/01/05260200. (put on the "to paragraph" page)
and then  Maybe make a "section" and then embed in 05260300

¨When a paragraph is used for positioning and the paragraph extends over more than one page or column, then positioning options that refer to a *paragraph area* or *paragraph text area* are applied only to the portion of the paragraph in the page or column where the anchor is located."

Change-Id: I8173050620fbeb8fc78d747ad05d8ab1f74f1346