tdf#126658 update "defining line ends"
The content of the guide page "Defining Line Ends"
(shared/guide/lineend_define.xhp) has not been updated since
its initial checkin in 2004. This patch updates the relevant
menu commands, aligns the description with the current UI,
adds a little more precision in what gets added as an arrowhead,
and refactor to <h1> and <menuitem>.

Also, add a <relatedtopics> section to the "Arrow Styles" help page
(shared/01/05200300.xhp) with an embedded link to this guide page.

Change-Id: I3e619aadddb65b77cc7216074d18c4eb138b1774