Drop this nonsense
Phrases like this look like an explanation form a clueless person.
The help should rely on basics (like differentiating text from names),
and not suggest to "place the number in quotation marks" when the
argument is documented to be a text. If needed, the word "text" may
link to a glossary article with syntax explanation.

Change-Id: I32e7a3c68659f74db6bc30c5a3fa9acdd158acb4
Reviewed-on: https://gerrit.libreoffice.org/c/help/+/136238
Tested-by: Jenkins
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>
(cherry picked from commit cc62afdf93da53b17e1b7859376d01b03292de9c)
Reviewed-on: https://gerrit.libreoffice.org/c/help/+/136424
Reviewed-by: Adolfo Jayme Barrientos <fitojb@ubuntu.com>