Numeral system -> Number system
Numeral refers to the symbols used in numbers: arabic, roman, ...

A number system is determined by the number of digits available for representing numbers.
The decimal system, for instance is based on the ten digits (0..9),
the binary system is based on the two digits 0 and 1,
the hexadecimal system is based on 16 digits (0...9 and A...F).

Change-Id: Ieef6581b04306b78b2cd5cd461ed0bd6b049fb07