tdf#150857 Better x-ref Table Formulas in Help
As requested in the bug report, this patch adds a bookmark so that searching for "table formulas" will be easier to find in the Help.

This patch also fixes inconsistencies as to how this feature is accessed in Writer. It is not necessary to have the cursor inside a table to be able to insert formulas in text documents.

Change-Id: Icb94193d2a274ff4e19255cb46c5fdad5cc3b2b3
Reviewed-on: https://gerrit.libreoffice.org/c/help/+/140170
Tested-by: Jenkins
Reviewed-by: Olivier Hallot <olivier.hallot@libreoffice.org>