Related tdf#149221 Create an entry for the Help button in Frequently-Used Buttons page
The help page "Frequently-Used Buttons" (text/shared/00/00000001.xhp) lacks an entry for the Help buttons in dialogs. This patch creates such entry.

This patch also adds a HID target so that when the user presses F1 and the Help button has focus in the Database Wizard, this page is then shown.

Other wizard dialogs should use a similar strategy when F1 is pressed and the Help button is focused.

Change-Id: I1295a7711c30e9cf47dbb18c784541e75eb90b71
Reviewed-on: https://gerrit.libreoffice.org/c/help/+/142915
Tested-by: Jenkins
Reviewed-by: Olivier Hallot <olivier.hallot@libreoffice.org>