cd to helpcontent2/source/text & drop it from filenames for larger batches
On the Jenkins Windows builder only 40~45 of the ~2500 helpfiles can be
processed at once when not stripping that prefix. With the shortened
filenames the build needs less than half of the xsltproc invocations.

Also prevent a trailing carriage return to be interpreted as part of the
filename on windows and drop the abbreviate_dir wrapper around the
command since there's nothing left to shorten anymore.

Change-Id: Ifaff26f73f262bd1d7cb4f813bf081782661ba17
Reviewed-on: https://gerrit.libreoffice.org/c/help/+/144236
Tested-by: Jenkins
Reviewed-by: Christian Lohmaier <lohmaier+LibreOffice@googlemail.com>
(cherry picked from commit 5daaa3e440b17505d644273a7305f073d6d5834d)
Reviewed-on: https://gerrit.libreoffice.org/c/help/+/144264