Related tdf#151309 Use t-test instead of T Test
As discussed in the ticket, we should use "t-test" instead of "T Test".

A previous patch fixed the occurrences in the UI. This patch fixes this only occurrence in the help files.

Actually the entire text uses "t-test" but the page title uses "T Test". So here I made the page title better match the title used in the heading <h1>.

Change-Id: I50d8814af5d4f93568aca7e5c299d95664bdb582
Reviewed-on: https://gerrit.libreoffice.org/c/help/+/149607
Tested-by: Jenkins
Reviewed-by: Olivier Hallot <olivier.hallot@libreoffice.org>