Make Help navigation layout less busy
Bookmarks don't have to be visible all the time, so make them a
dropdown and move the search box to header.

Move Xapian search to the top of the content area.

Make Contents tree appear on the left side in 960px and wider screens.

Simplify/clean up some CSS and JS, for example for module colours.

Change-Id: Ic50a834b758d50098868866e4b8d3e8085755423
Reviewed-on: https://gerrit.libreoffice.org/c/help/+/160870
Tested-by: Jenkins
Reviewed-by: Olivier Hallot <olivier.hallot@libreoffice.org>