tdf#155877 Fix broken embed links in Help pages
Created a section called "methods_toc" in source/text/sbasic/shared/03/sf_writer.xhp so that the Methods section is embedded correctly in other files.

Replaced link in text/scalc/guide/table_cellmerge.xhp to "text/shared/01/05100200.xhp#unmergecells"  with a link to "text/shared/01/05100200.xhp#h1" so that it links to the correct content for Unmerge Cells

Replaced link in /text/scalc/main0218.xhp to "text/scalc/00/00000405.xhp#syautoformat" with a link to "text/shared/00/00040503.xhp#syautoformat" so that the help article displays the correct howtoget information for Autoformat Styles.

Replaced link in /text/scalc/main0218.xhp to "text/scalc/00/00000405.xhp#syautoformat" with a link to "text/shared/00/00040503.xhp#syautoformat" so that the help article displays the correct howtoget information for Autoformat Styles.

Replaced link in /text/scalc/main0218.xhp to "text/scalc/02/06080000.xhp#sythemes" with a link to "text/scalc/00/00000405.xhp#sythemes" so that the help article displays the correct howtoget information for Choose Themes.

Replaced link in /text/shared/main0201.xhp to "text/shared/01/03010000.xhp#zoom" with a link to "text/shared/01/03010000.xhp#zoomlayout" so that it displays the correct help page information for "Zoom & View Layout".

Corrected typo in text/shared/main0204.xhp so that the embedded link to source/text/swriter/00/00000405.xhp displays the correct information for Delete Column.

Removed link in /text/swriter/main0204.xhp to "source/text/shared/01/AutoFormat.xhp#autoformattabelle" and replaced it with links to "text/shared/01/AutoFormat.xhp#par_id3145367" and "text/shared/00/00040503.xhp#syautoformat", so that it displays the correct description and icon for Autoformat.

Change-Id: I261d19e0c819ac8bf2f7f83e6e540f76e5eabacd
Reviewed-on: https://gerrit.libreoffice.org/c/help/+/163018
Tested-by: Jenkins
Reviewed-by: Olivier Hallot <olivier.hallot@libreoffice.org>