#tdf120179 Rewrite SUMIF Help Page
Created a new file named text/scalc/01/func_sumif.xhp. This file includes updated help information for the SUMIF function, including a more precise definitiion of and more detailed examples.

Added the new file text/scalc/01/func_sumif.xhp to AllLangHelp_scalc.mk so that it will be included in the Help application.

Replaced the content in text/scalc/01/04060106#Section15 ("SUMIF") with an embed link to the new text/scalc/01/func_sumif.xhp, so that only the definition and a link to the main SUMIF Function page appears in the list of Mathematical Functions.

Updated the "Related Topics" sections of text/scalc/01/func_sum.xhp and text/scalc/01/func_sumifs.xhp so that they link to the new SUMIF Function page.

Change-Id: I224dffb0c3c6411ea416bba308559f9fe3896dd2
Reviewed-on: https://gerrit.libreoffice.org/c/help/+/165186
Tested-by: Jenkins
Reviewed-by: Olivier Hallot <olivier.hallot@libreoffice.org>