tdf#159872 Review Help XLOOKUP XMATCH functions
From comments in by Regina and Ady in
https://community.documentfoundation.org/t/comments-on-help-for-new-xlookup-fucntion/12086?u=ohallot
and following.

+ Note on "case sensitive"
+ Details on array and range
+ Fix #N/A to "returns the #N/A error"
+ Reworked examples (re-use HLOOKUP table)
+ Grammar and linguistics here and there
+ Fix syntax definition
+ Fix "Not found" definition
+ Fix MatchType using verb "attempt"
+ Position of optional arguments
+ Longer user cases to be addressed in wiki

Note: the function clones Excel XLOOKUP but it is not a copy. Therefore
the name of the arguments were set to differentiate from Excel. The
Calc function arguments usual naming were mostly mnemonic and
not verbose. The current XLOOKUP names in the function wizard is
verbose and has several words. Review is needed.

Patch set 2: fix xml DTD
Patch set 3: XMATCH
+ Reworked example (re-use HLOOKUP table)
+ Link to MATCH
+ Fix function syntax
+ Refactor
Patch set 4:
+ typos
+ Fix case insensitive

Change-Id: I9f085dcadbdb013cf3a3b892791e18c729aa6431
Reviewed-on: https://gerrit.libreoffice.org/c/help/+/168072
Tested-by: Jenkins
Reviewed-by: Olivier Hallot <olivier.hallot@libreoffice.org>