tdf#162504 Create help page for Number Format pane
Create new page source/text/scalc/sidebar_number_format.xhp to describe
the features of the Number Format pane in the Properties sidebar deck.

In AllLangHelp_scalc.mk
- Add new file source/text/scalc/sidebar_number_format.xhp to
repository.

In source/text/scalc/01/default_number_formats.xhp
- Add variable and section tags so that content can be easily linked and
embedded in other pages.
- Remove old comments

In source/text/shared/01/05020300.xhp
- Add variable and section tags so that content can be easily linked and
embedded in other pages.
- Remove old comments

In
/home/dione/libreoffice/helpcontent2/source/text/shared/01/sidebar_properties.xhp
THIS PAGE IS A WIP AND WILL NOT BE LINKED IN HELP UNTIL IT IS FINISHED
- Embed description of the help page for Number Formats
- Added index bookmarks (Currently commented out)
- Updated application switches to hide the desription of the Paragraph
pane for CALC and BASE.

Change-Id: I7e1849ddc41cad7fd5d4b6c57e8a03186eb6575f
Reviewed-on: https://gerrit.libreoffice.org/c/help/+/176720
Tested-by: Jenkins
Tested-by: Olivier Hallot <olivier.hallot@libreoffice.org>
Reviewed-by: Olivier Hallot <olivier.hallot@libreoffice.org>